# swagger/delilah_resto.yaml
...
swagger: "2.0"
info:
  description: "This is the OpenAPI specification for the Delilah Restó system"
  version: "1.0.0"
  title: "Delilah Restó"
  contact:
    email: "ibrian_84@hotmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/IBRIAN/delilah_resto/1.0.0"
...
schemes:
- "https"
paths:
...
  /users/{username}:
    get:
      tags:
      - "user"
      summary: "Fetches an specific user"
      description: "A determined user can be fetched as well as its complete information.\
        \ Only users with an \"Administrator\" role can fetch users."
      operationId: "fetchUser"
      parameters:
      - name: "username"
        in: "path"
        description: "The desired user's username"
        required: true
        type: "string"
      - name: "authorization"
        in: "header"
        description: "Bearer Type JSON Web Token"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  User:
    type: "object"
    required:
    - "userAddress"
    - "userEmail"
    - "userFirstName"
    - "userLastName"
    - "userPassword"
    - "userPhoneNumber"
    - "userRole"
    - "username"
    properties:
      userFirstName:
        type: "string"
        example: "Ibrian"
      userLastName:
        type: "string"
        example: "Festorazzi"
      userEmail:
        type: "string"
        format: "email"
        example: "ibrian_84@hotmail.com"
      userPhoneNumber:
        type: "string"
        example: "123939938"
      userAddress:
        type: "string"
        example: "m89 c31"
      username:
        type: "string"
        example: "ibra"
      userPassword:
        type: "string"
        format: "password"
        example: "1234"
      userRole:
        type: "string"
        example: "Administrator"
...
  ErrorResponse:
    type: "object"
    properties:
      errordescription:
        type: "string"
        example: "Unable to create user"
      errorDetail:
        type: "string"
        example: "There is already an user registered with this email"